fix Makefile
