+++ lib/es6/caml_format.js
@@ -265,12 +265,10 @@ function caml_int64_of_string(s) {
     };
   };
   var res = Caml_int64.mul(sign, aux(d, i + 1 | 0));
-  var or_res_000 = /* hi */res[0] | /* hi */0;
-  var or_res_001 = /* lo */(res[1] >>> 0);
-  var or_res = /* int64 */[
-    or_res_000,
-    or_res_001
-  ];
+  var or_res = Caml_int64.or_(res, /* int64 */[
+        /* hi */0,
+        /* lo */0
+      ]);
   if (Caml_int64.eq(base, /* int64 */[
           /* hi */0,
           /* lo */10
