+++ lib/es6/caml_int64.js
@@ -292,6 +292,27 @@ function swap(param) {
         ];
 }
 
+function xor(param, param$1) {
+  return /* record */[
+          /* hi */param[/* hi */0] ^ param$1[/* hi */0],
+          /* lo */((param[/* lo */1] ^ param$1[/* lo */1]) >>> 0)
+        ];
+}
+
+function or_(param, param$1) {
+  return /* record */[
+          /* hi */param[/* hi */0] | param$1[/* hi */0],
+          /* lo */((param[/* lo */1] | param$1[/* lo */1]) >>> 0)
+        ];
+}
+
+function and_(param, param$1) {
+  return /* record */[
+          /* hi */param[/* hi */0] & param$1[/* hi */0],
+          /* lo */((param[/* lo */1] & param$1[/* lo */1]) >>> 0)
+        ];
+}
+
 function ge(param, param$1) {
   var other_hi = param$1[/* hi */0];
   var hi = param[/* hi */0];
@@ -609,6 +630,9 @@ export{
   asr_          ,
   is_zero       ,
   mul           ,
+  xor           ,
+  or_           ,
+  and_          ,
   swap          ,
   ge            ,
   eq            ,
