+++ lib/es6/caml_oo.js
@@ -3,8 +3,7 @@
 import * as Caml_array              from "./caml_array";
 import * as Caml_builtin_exceptions from "./caml_builtin_exceptions";
 
- var caml_methods_cache = Caml_array.caml_make_vect(1000, 0);
-
+var caml_methods_cache = Caml_array.caml_make_vect(1000, 0);
 
 function caml_get_public_method(obj, tag, cacheid) {
   var meths = obj[0];
