+++ lib/es6/curry.js
@@ -1,6 +1,5 @@
 'use strict';
 
-import * as Caml_oo    from "./caml_oo";
 import * as Caml_array from "./caml_array";
 
 function app(_f, _args) {
@@ -30,11 +29,6 @@ function app(_f, _args) {
       return f.apply(null, args);
     }
   };
-}
-
-function js(label, cacheid, obj, args) {
-  var meth = Caml_oo.caml_get_public_method(obj, label, cacheid);
-  return app(meth, args);
 }
 
 function curry_1(o, a0, arity) {
@@ -83,10 +77,6 @@ function _1(o, a0) {
   else {
     return curry_1(o, a0, arity);
   }
-}
-
-function js1(label, cacheid, a0) {
-  return _1(Caml_oo.caml_get_public_method(a0, label, cacheid), a0);
 }
 
 function __1(o) {
@@ -150,10 +140,6 @@ function _2(o, a0, a1) {
   }
 }
 
-function js2(label, cacheid, a0, a1) {
-  return _2(Caml_oo.caml_get_public_method(a0, label, cacheid), a0, a1);
-}
-
 function __2(o) {
   var arity = o.length;
   if (arity === 2) {
@@ -215,10 +201,6 @@ function _3(o, a0, a1, a2) {
   else {
     return curry_3(o, a0, a1, a2, arity);
   }
-}
-
-function js3(label, cacheid, a0, a1, a2) {
-  return _3(Caml_oo.caml_get_public_method(a0, label, cacheid), a0, a1, a2);
 }
 
 function __3(o) {
@@ -285,10 +267,6 @@ function _4(o, a0, a1, a2, a3) {
   else {
     return curry_4(o, a0, a1, a2, a3, arity);
   }
-}
-
-function js4(label, cacheid, a0, a1, a2, a3) {
-  return _4(Caml_oo.caml_get_public_method(a0, label, cacheid), a0, a1, a2, a3);
 }
 
 function __4(o) {
@@ -359,10 +337,6 @@ function _5(o, a0, a1, a2, a3, a4) {
   else {
     return curry_5(o, a0, a1, a2, a3, a4, arity);
   }
-}
-
-function js5(label, cacheid, a0, a1, a2, a3, a4) {
-  return _5(Caml_oo.caml_get_public_method(a0, label, cacheid), a0, a1, a2, a3, a4);
 }
 
 function __5(o) {
@@ -438,10 +412,6 @@ function _6(o, a0, a1, a2, a3, a4, a5) {
   else {
     return curry_6(o, a0, a1, a2, a3, a4, a5, arity);
   }
-}
-
-function js6(label, cacheid, a0, a1, a2, a3, a4, a5) {
-  return _6(Caml_oo.caml_get_public_method(a0, label, cacheid), a0, a1, a2, a3, a4, a5);
 }
 
 function __6(o) {
@@ -523,10 +493,6 @@ function _7(o, a0, a1, a2, a3, a4, a5, a6) {
   else {
     return curry_7(o, a0, a1, a2, a3, a4, a5, a6, arity);
   }
-}
-
-function js7(label, cacheid, a0, a1, a2, a3, a4, a5, a6) {
-  return _7(Caml_oo.caml_get_public_method(a0, label, cacheid), a0, a1, a2, a3, a4, a5, a6);
 }
 
 function __7(o) {
@@ -619,10 +585,6 @@ function _8(o, a0, a1, a2, a3, a4, a5, a6, a7) {
   }
 }
 
-function js8(label, cacheid, a0, a1, a2, a3, a4, a5, a6, a7) {
-  return _8(Caml_oo.caml_get_public_method(a0, label, cacheid), a0, a1, a2, a3, a4, a5, a6, a7);
-}
-
 function __8(o) {
   var arity = o.length;
   if (arity === 8) {
@@ -637,38 +599,29 @@ function __8(o) {
 
 export{
   app     ,
-  js      ,
   curry_1 ,
   _1      ,
-  js1     ,
   __1     ,
   curry_2 ,
   _2      ,
-  js2     ,
   __2     ,
   curry_3 ,
   _3      ,
-  js3     ,
   __3     ,
   curry_4 ,
   _4      ,
-  js4     ,
   __4     ,
   curry_5 ,
   _5      ,
-  js5     ,
   __5     ,
   curry_6 ,
   _6      ,
-  js6     ,
   __6     ,
   curry_7 ,
   _7      ,
-  js7     ,
   __7     ,
   curry_8 ,
   _8      ,
-  js8     ,
   __8     ,
   
 }
