+++ lib/es6/int64.js
@@ -30,10 +30,10 @@ function abs(n) {
 }
 
 function lognot(n) {
-  return /* int64 */[
-          /* hi */n[0] ^ /* hi */-1,
-          /* lo */((n[1] ^ /* lo */4294967295) >>> 0)
-        ];
+  return Caml_int64.xor(n, /* int64 */[
+              /* hi */-1,
+              /* lo */4294967295
+            ]);
 }
 
 function to_string(n) {
