+++ lib/es6/random.js
@@ -140,12 +140,10 @@ function int64(s, bound) {
       var b1 = Caml_int64.of_int32(bits(s$1));
       var b2 = Caml_int64.lsl_(Caml_int64.of_int32(bits(s$1)), 30);
       var b3 = Caml_int64.lsl_(Caml_int64.of_int32(bits(s$1) & 7), 60);
-      var r_000 = /* hi */b1[0] | /* hi */b2[0] | b3[0];
-      var r_001 = /* lo */((b1[1] | b2[1] | b3[1]) >>> 0);
-      var r = /* int64 */[
-        r_000,
-        r_001
-      ];
+      var r = Caml_int64.or_(b1, /* int64 */[
+            /* hi */b2[0] | b3[0],
+            /* lo */((b2[1] | b3[1]) >>> 0)
+          ]);
       var v = Caml_int64.mod_(r, n);
       if (Caml_int64.gt(Caml_int64.sub(r, v), Caml_int64.add(Caml_int64.sub(Int64.max_int, n), /* int64 */[
                   /* hi */0,
